% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMatch.R
\name{predictMatch}
\alias{predictMatch}
\title{Predict Match}
\usage{
predictMatch(homeTeam, awayTeam, neutralLocation = FALSE)
}
\arguments{
\item{homeTeam}{a `team` object of the home team (or just team one if played at a neutral location).}

\item{awayTeam}{a `team` object of the away team (or just team two if played at a neutral location).}

\item{neutralLocation}{a boolean, True if the match to be pedicted will not be played at the home stadium of either team.}
}
\value{
a list of length two the first element, `teamOnePts`, is the predicted amount of points for the first team. `teamTwoPts` is the predicted amount for the second team.
}
\description{
Prints out a predicted score based on prior performance of the teams.
}
\examples{
CFB<-scrapeCurrentSeason()
OU<-team("Oklahoma", CFB)
OSU<-team("Ohio State", CFB)
predictMatch(Oklahoma, Ohio State)

}
