% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeCurrentSeason.R
\name{scrapeCurrentSeason}
\alias{scrapeCurrentSeason}
\title{scrapeCurrentSeason}
\usage{
scrapeCurrentSeason()
}
\value{
a dataframe of up to date information, in the same form as the CFB19wk13 dataframe
}
\description{
scrapes the web to get the most up to date rankings and scores of the current college football season for all 130 CFB teams
}
\examples{
CFB.df<-scrapeCurrentSeason()

}
