% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cfbTeam.R
\name{summary.cfbTeam}
\alias{summary.cfbTeam}
\title{Summary}
\usage{
\method{summary}{cfbTeam}(team)
}
\arguments{
\item{team}{team object to be summarized}
}
\value{
a list of the teams current record, reported as "Nwins-Nlosses", the teams CBS rank and the
defense and offense scores assigned by odds shark.
}
\description{
Summary function for cfbTeam object.
}
\examples{
CFB<-data(CFB19wk13)
OSU<-team("Ohio State", CFB)
summary(OSU)
}
