% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{team}
\alias{team}
\title{cfbTeam constructor}
\usage{
team(SpRefTeamName, season.df)
}
\arguments{
\item{SpRefTeamName}{The name of the team according to sports reference}

\item{season.df}{The season data from which the team data will be extracted}
}
\value{
a cfbTeam object
}
\description{
Constructor function for the cfbTeam object which encompasses the data for an FBS team useable in this package.
}
\examples{
CFB<-data(CFB19wk13)
OSU<-team("Ohio State", CFB)

}
