% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamNames.R
\docType{data}
\name{teamNames}
\alias{teamNames}
\title{teamNames}
\format{A data frame containing 130 observations of 3 variables
\describe{
\item{CBS}{The name of the team according to CBS}
\item{SpRef}{The name of the team according to sports reference}
\item{OddShark}{The name of the team according to OddsShark}
}}
\source{
\url{https://www.sports-reference.com/cfb/years/2019-schedule.html},\url{https://www.cbssports.com/college-football/rankings/cbs-sports-ranking/},\url{https://www.oddsshark.com/ncaaf/defensive-stats}
}
\usage{
teamNames
}
\description{
A data set containing every team name both from CBS sports and Sports Reference. To be used by the package to resolve discrepencies between the two sources
}
\keyword{datasets}
