% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validTeamNames.R
\name{validTeamNames}
\alias{validTeamNames}
\title{Valid team names}
\usage{
validTeamNames()
}
\value{
a character vector of all valid team names
}
\description{
A method to see all of FBS team names according to sports reference
}
\examples{
validTeamNames()
}
